% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_bad.R
\name{adjust_points2}
\alias{adjust_points2}
\title{Adjust points but with NAs instead of last row.}
\usage{
adjust_points2(data, npoints)
}
\arguments{
\item{data}{The dataset}

\item{npoints}{The correct number of points in the cap}
}
\value{
the dataset with the final number of points
}
\description{
Adjust points but with NAs instead of last row.
}
