% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{align_all_caps_nested}
\alias{align_all_caps_nested}
\title{Align all nested caps to all other caps in a cap size}
\usage{
align_all_caps_nested(data, num_aligned)
}
\arguments{
\item{data}{the digitizeR nested data list}

\item{num_aligned}{the number of points you want to use to do the alignment}
}
\value{
a list of lists which contains every participants rotation to every other participant within each cap size
}
\description{
Align all nested caps to all other caps in a cap size
}
\details{
use the modified kabsch to align every cap to every other cap in a certain size.
Fairly intensive for large data sets so use wisely.
}
