% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_kabsch.R
\name{align_to_template}
\alias{align_to_template}
\title{Align a dataset to the template for each cap size}
\usage{
align_to_template(template, data, num_aligned)
}
\arguments{
\item{template}{The template dataset}

\item{data}{the data output of select_caps_by_npoints}

\item{num_aligned}{the number of points to use in the alignment}
}
\value{
the listed dataset aligned to the template in each cap size
}
\description{
This is done only for a full set of alignments, in other occasions you can use align_to_template2
}
