% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_kabsch.R
\name{align_to_template2}
\alias{align_to_template2}
\title{Align a partial dataset to the template for each cap size}
\usage{
align_to_template2(template, data, num_aligned)
}
\arguments{
\item{template}{The template dataset}

\item{data}{the data output of select_caps_by_npoints}

\item{num_aligned}{the number of points to use in the alignment}
}
\value{
the listed dataset aligned to the template in each cap size
}
\description{
This is used for a partial set of alignments if needed.
}
