% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_3d_dist.R
\name{calc_3d_dist}
\alias{calc_3d_dist}
\title{calculate 3d distance}
\usage{
calc_3d_dist(point0, point1)
}
\arguments{
\item{point0}{A point x y z}

\item{point1}{A point x y z}
}
\value{
The 3d distance from 0 to 1
}
\description{
A helper function for 3d distance
}
