% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_3d_dist.R
\name{calc_dist_and_replace_na}
\alias{calc_dist_and_replace_na}
\title{calculate the 3d distance and replace with NAs when outside the specified distance}
\usage{
calc_dist_and_replace_na(template, aligned_data, original_data, max)
}
\arguments{
\item{template}{the templates to align to}

\item{aligned_data}{the data post-alignment}

\item{original_data}{the data before alignment}

\item{max}{the maximum distance}
}
\value{
Aligned data with NAs replacing the points that are out of bounds
}
\description{
calculate the 3d distance and replace with NAs when outside the specified distance
}
