% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_3d_dist.R
\name{calc_participant_numbers}
\alias{calc_participant_numbers}
\title{Calculate participant numbers}
\usage{
calc_participant_numbers(data)
}
\arguments{
\item{data}{the data set}
}
\value{
the number of participants in the dataset
}
\description{
Calculate participant numbers
}
