% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{count_nested_aligned_caps}
\alias{count_nested_aligned_caps}
\title{count all caps that align with each participant in each cap size}
\usage{
count_nested_aligned_caps(nested_data, original_data, max, points)
}
\arguments{
\item{nested_data}{the dataset which results from align_all_caps_nested}

\item{original_data}{The listed caps after using select_caps_by_npoints}

\item{max}{The maximum acceptable distance in cm}

\item{points}{The number of points that were used in alignment previously}
}
\value{
a list of participants who align with any other participant
}
\description{
count all caps that align with each participant in each cap size
}
\details{
This works on the basis of a maximum acceptable distance given by max.
}
