% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_bad.R
\name{find_greatest_alignment}
\alias{find_greatest_alignment}
\title{Find the best cap to align to}
\usage{
find_greatest_alignment(aligned_caps)
}
\arguments{
\item{aligned_caps}{the nested list from count_nested_aligned_caps}
}
\value{
A list of best candidates in each cap size
}
\description{
This is a helpful step to make the templates which allows visualisation of whether a sensible choice has been made by the algorithm.
}
