% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_bad.R
\name{greatest_alignment}
\alias{greatest_alignment}
\title{See the number of caps that align with the best candidate}
\usage{
greatest_alignment(aligned_caps)
}
\arguments{
\item{aligned_caps}{the nested list from count_nested_aligned_caps}
}
\value{
The number of caps the best candidate aligns with in each cap size
}
\description{
This function gets used to check that the best candidate is aligning to enough other caps to be an acceptable choice. If it is too low, the distance should be increased above.
}
