% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_3d_dist.R
\name{headwise_replacement}
\alias{headwise_replacement}
\title{Headwise replacement}
\usage{
headwise_replacement(template, data, max)
}
\arguments{
\item{template}{the template dataset}

\item{data}{the dataset to replace}

\item{max}{the maximum distance to accept}
}
\value{
The dataset with caps replaced with template caps when poor
}
\description{
replaces by a whole cap rather than individual points.
When the average distance for the cap goes beyond the max value, the whole cap is replaced
}
