% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_3d_dist.R
\name{iterative_replacement}
\alias{iterative_replacement}
\title{iterative replacement}
\usage{
iterative_replacement(
  template,
  aligned_data,
  original_data,
  npoints_in_cap,
  max_threshold
)
}
\arguments{
\item{template}{the template dataset}

\item{aligned_data}{the data after alignment}

\item{original_data}{the original dataset}

\item{npoints_n_cap}{the number of points in the cap.}
}
\value{
Data aligned to the template in an iterative fashion
}
\description{
Iteratively replaces with nas until the end when replaced by template.
Reduced by integer until below 5, when it is .5
}
