% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_templates.R
\name{make_full_templates}
\alias{make_full_templates}
\title{Make final or full templates after the partial templates}
\usage{
make_full_templates(data, template_index, distances, length)
}
\arguments{
\item{data}{the dataset. In normal circumstances should be the output of align_to_template to the partial template.}

\item{template_index}{Output of find_greatest_alignment}

\item{distances}{Output of count_nested_aligned_caps}

\item{length}{The number of points that exist in the cap}
}
\value{
a template in each cap size
}
\description{
The expectation is that this is used after making partial templates and aligning to them. This then takes the output of those functions
}
