% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_templates.R
\name{make_partial_templates}
\alias{make_partial_templates}
\title{create partial preliminary templates from data}
\usage{
make_partial_templates(template_index, data)
}
\arguments{
\item{template_index}{output of find_greatest_alignment}

\item{data}{raw data output of select_caps_by_npoints}
}
\value{
one sample cap for each cap size
}
\description{
This is used as a first step before creating the final template.
The idea is just to check we are on the right track.
}
