% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modified_kabsch.R
\name{modified_kabsch2}
\alias{modified_kabsch2}
\title{a second modified version of the kabsch algorithm to work on a subset of a matrix.}
\usage{
modified_kabsch2(pm, qm, om)
}
\arguments{
\item{pm}{a matrix}

\item{qm}{a matrix}

\item{om}{a third matrix}
}
\value{
a transposed om
}
\description{
a second modified version of the kabsch algorithm to work on a subset of a matrix.
}
