% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_one_cap.R
\name{plot_one_participant}
\alias{plot_one_participant}
\title{Plot the cap of a single participant}
\usage{
plot_one_participant(data, index, part, flipped = T)
}
\arguments{
\item{data}{the dataset from which to plot}

\item{index}{an integer indicating the ordinal capsize list you wish to choose from}

\item{part}{an integer indicating the ordinal participant number you wish to choose in that cap size}

\item{flipped}{logical indicating whether the cap should be flipped on the z axis for visualisation}
}
\value{
a 3d plot of the points of one cap
}
\description{
Plot the cap of a single participant
}
\details{
This is done by numbers, so you choose to plot eg fro the second cap size the fourth participant.
}
