% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_one_cap.R
\name{plot_template}
\alias{plot_template}
\title{Plot a template from the template dataset to see how they look}
\usage{
plot_template(template, index, flipped = T)
}
\arguments{
\item{template}{the tamplate dataset}

\item{index}{the numerical index of which cap size to choose}

\item{flipped}{logical, if true will diplay cap right way up}
}
\value{
a full plot
}
\description{
This is a way to plot all the points in a template to see the spread of digipoints.
}
