% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{prepare_and_make_templates}
\alias{prepare_and_make_templates}
\title{prepare and make template}
\usage{
prepare_and_make_templates(original_data, permitted_dist, npoints)
}
\arguments{
\item{original_data}{the original dataset after selecting by number of points}

\item{permitted_dist}{the distance within which to allow alignment. If unacceptable try again}

\item{npoints}{the number of points in the cap}
}
\value{
The final templates
}
\description{
A wrapper function to do the pre-alignment, distance checks, and template creation needed.
This can be slow for a large dataset!
}
