% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_caps.R
\name{read_in_caps}
\alias{read_in_caps}
\title{read in caps}
\usage{
read_in_caps(digipath, pattern, capsize, type = c("NIHVWM", "Gates", "HWB"))
}
\arguments{
\item{digipath}{path to digitizations}

\item{pattern}{a pattern in files passed to list.files}

\item{capsize}{the path to the cap size document}

\item{type}{to be adjusted for cap geometry}
}
\value{
a list of digitizations in dataframe form
}
\description{
read in caps
}
\details{
currently in beta form for generalisation to other projects. For argument "type", running this depends on locations of short sources and layout of sources vs detectors.
}
