% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_caps.R
\name{read_in_templates}
\alias{read_in_templates}
\title{read in templates}
\usage{
read_in_templates(path)
}
\arguments{
\item{path}{the path to the templates}

\item{digipath}{path to digitizations}
}
\value{
a list of digitizations in dataframe form
}
\description{
read in templates
}
\details{
If templates have been previously made, they can be loaded in with this function.
}
