% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wpd.R
\name{read_wpd}
\alias{read_wpd}
\title{Read WebPlotDigitizer JSON data}
\usage{
read_wpd(txt, include_pixel = FALSE)
}
\arguments{
\item{txt}{Input txt as readable by fromJSON}

\item{include_pixel}{Include the pixel data in the returned data frame}
}
\value{
A data.frame with colums for \code{x}, \code{y}, and \code{DataSet}
  and optionally (if \code{include_pixel=TRUE}) \code{x_pixel} and
  \code{y_pixel}.
}
\description{
Read WebPlotDigitizer JSON data
}
