% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_3d_dist.R
\name{replace_nas_with_template}
\alias{replace_nas_with_template}
\title{replace NAs with template}
\usage{
replace_nas_with_template(template, data)
}
\arguments{
\item{template}{template data}

\item{data}{the cap dataset}
}
\value{
the data with NAs replaced with values from the template
}
\description{
A simple replacement function for NAs
}
