% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{rezero_caps}
\alias{rezero_caps}
\title{rezero caps}
\usage{
rezero_caps(data)
}
\arguments{
\item{data}{the data set}
}
\value{
the data with each cap rezeroed so that CZ is at 0
}
\description{
rezero caps
}
