% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_templates.R
\name{rezero_template}
\alias{rezero_template}
\title{Rezero template}
\usage{
rezero_template(template)
}
\arguments{
\item{template}{The template dataset (usually the output of make_full_templates)}
}
\value{
Slightly cleaned up version of templates
}
\description{
This is a helper function to allow bringing the z value of the cz point back to zero so you are largely in the same space. It also sanity checks for outlier points.
}
