% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_templates.R
\name{rezero_template_ind}
\alias{rezero_template_ind}
\title{rezero and sanity check templates}
\usage{
rezero_template_ind(template)
}
\arguments{
\item{template}{the template dataset}
}
\value{
templates such that the CZ is at 0, and wide points are moved in
}
\description{
Contains extra sanity checks that are more conservative than rezero_template
}
