% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_caps.R
\name{save_caps}
\alias{save_caps}
\title{Save caps to disk as digpts.txt}
\usage{
save_caps(data, path, type = c("NIHVWM", "Gates", "HWB"))
}
\arguments{
\item{data}{data to save}

\item{path}{path to save the data}

\item{type}{to save the data for a specific cap geometry.}
}
\value{
Saved digpts.txt to the path in a folder for each participant, ready to use in AtlasViewerGUI
}
\description{
This is useful if you want to use individual caps for each participant. Experimental - the type argument would need adjustment for a specific cap geometry.
}
