% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{select_caps_by_npoints}
\alias{select_caps_by_npoints}
\title{select caps to use by a certain number of points}
\usage{
select_caps_by_npoints(data, npoints)
}
\arguments{
\item{data}{the data in list format from read_in_caps()}

\item{npoints}{the correct number of points without which the digi is invalid.}
}
\value{
the data list containing only the valid digis
}
\description{
select caps to use by a certain number of points
}
\details{
take any number of nested caps and select only the ones with the right number of points.
}
