% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_functions.R
\name{select_caps_without_npoints}
\alias{select_caps_without_npoints}
\title{select only the caps without a number of points}
\usage{
select_caps_without_npoints(data, npoints)
}
\arguments{
\item{data}{the overall dataset to pull from}

\item{npoints}{the number of points you want the data to NOT have}
}
\value{
the subset of data that misses points
}
\description{
this is useful on the occasions when you want to detail with caps that are obviously wrong because they miss a number of points.
}
