% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_3d_dist.R
\name{threestep_alignment}
\alias{threestep_alignment}
\title{three step replacement}
\usage{
threestep_alignment(
  template,
  aligned_data,
  original_data,
  npoints,
  dist1,
  dist2,
  dist3
)
}
\arguments{
\item{template}{The template dataset to align to}

\item{aligned_data}{the data in a preliminary aligned form}

\item{original_data}{the data after selecting by npoints}

\item{npoints}{The number of points in the cap (to align by)}

\item{dist1}{Numeric. The first distance to align at}

\item{dist2}{Numeric. The second distance to align at}

\item{dist3}{Numeric. The third distance to align at}
}
\value{
a dataset that has been realigned three times at different distances to replace hopefully no more than the needed number of points only.
}
\description{
This function is Experimental. Replaces points greater than the acceptable distance 3 times and replaces with NA to realign and try again.
At the end replaces with template data where outside of the required distance.
}
