% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digitizeR.R
\name{wpd.close}
\alias{wpd.close}
\title{Shutdown the HTTP server that is currently hosting WebPlotDigitizer.}
\usage{
wpd.close(app)
}
\arguments{
\item{app}{Server handle that was obtained by executing wpd.launch()}
}
\description{
Shutdown the HTTP server that is currently hosting WebPlotDigitizer.
}
\examples{
wpd.close(app)
}
