% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digitizeR.R
\name{wpd.isOpen}
\alias{wpd.isOpen}
\title{Check if the HTTP server is currently running.}
\usage{
wpd.isOpen(app)
}
\arguments{
\item{app}{Server handle that was obtained by executng wpd.launch()}
}
\description{
Check if the HTTP server is currently running.
}
