% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/digitizeR.R
\name{wpd.launch}
\alias{wpd.launch}
\title{Start a local HTTP server that hosts WebPlotDigitizer.
This will also open a browser window pointing to the local URL.}
\usage{
wpd.launch(location = "0.0.0.0", port = 8000)
}
\arguments{
\item{location}{IP address or machine name of the server. Defaults to "0.0.0.0".}

\item{port}{Port number of the HTTP server. Defaults to 8000.}
}
\value{
Server handle that is later used to shutdown the server using wpd.close()
}
\description{
Start a local HTTP server that hosts WebPlotDigitizer.
This will also open a browser window pointing to the local URL.
}
\examples{
app = wpd.launch()
app = wpd.launch(port=8080)
app = wpd.launch(location="192.168.1.100", port=8080)
}

