% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digitizeR.R
\name{wpd_close}
\alias{wpd_close}
\title{Shutdown the HTTP server that is currently hosting WebPlotDigitizer.}
\usage{
wpd_close(app)
}
\arguments{
\item{app}{Server handle that was obtained by executing wpd_launch()}
}
\description{
Shutdown the HTTP server that is currently hosting WebPlotDigitizer.
}
\examples{
\dontrun{
wpd_close(app)
}
}
