% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digitizeR.R
\name{wpd_isOpen}
\alias{wpd_isOpen}
\title{Check if the HTTP server is currently running.}
\usage{
wpd_isOpen(app)
}
\arguments{
\item{app}{Server handle that was obtained by executng wpd_launch()}
}
\description{
Check if the HTTP server is currently running.
}
