% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_in_caps.R
\name{read_in_caps}
\alias{read_in_caps}
\title{read in caps}
\usage{
read_in_caps(
  digipath,
  pattern,
  capsize,
  landmarks = c("nz", "ar", "al", "cz", "iz"),
  num_source = 16,
  num_detector = 32,
  short = c(3, 11, 19, 27),
  subchar = c(1, 12)
)
}
\arguments{
\item{digipath}{path to digitizations as a string}

\item{pattern}{a pattern in files passed to list.files as a string}

\item{capsize}{the path to the cap size document as a string}

\item{landmarks}{the refpoints to be adjusted for cap geometry as a vector}

\item{num_source}{numeric value of the number of sources}

\item{num_detector}{numeric value of the number of detectors}

\item{short}{a vector indicating the numeric value of short sources, if any}

\item{subchar}{a vector indicating the characters from the ID to use, passed to substr}
}
\value{
a list of digitizations in dataframe form
}
\description{
read in caps
}
\details{
currently in beta form for generalisation to other projects. For argument "type", running this depends on locations of short sources and layout of sources vs detectors.
}
